requirements.txt
torch~=1.10.2
numpy~=1.20.3
tqdm~=4.62.3
torchvision~=0.11.3
pillow~=8.4.0
sklearn~=0.0
scikit-learn~=0.24.2
transformers~=4.20.1